feat:    * Dynamic Status Description    * Get/Registry configuration capabilities fix:    * A lot of bugs

+++ bundles/org.openhab.binding.lgthinq/src/main/java/org/openhab/binding/lgthinq/lgapi/model/ACSnapShotV1.java
@@ -23,7 +23,7 @@ public class ACSnapShotV1 extends ACSnapShot {
 
     @Override
     @JsonProperty("WindStrength")
-    public Double getAirWindStrength() {
+    public Integer getAirWindStrength() {
         return super.getAirWindStrength();
     }
 
@@ -41,13 +41,13 @@ public class ACSnapShotV1 extends ACSnapShot {
 
     @Override
     @JsonProperty("OpMode")
-    public Double getOperationMode() {
+    public Integer getOperationMode() {
         return super.getOperationMode();
     }
 
     @Override
     @JsonProperty("Operation")
-    public Double getOperation() {
+    public Integer getOperation() {
         return super.getOperation();
     }
 }
